+++ src/Resources/app/administration/src/component/nosto-job-info/component.js
@@ -7,6 +7,7 @@ import template from './nosto-job-info.html.twig';
 const { Mixin } = Shopware;
 const { Criteria } = Shopware.Data;
 
+/** @private */
 export default {
     template,
